# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: qa-automation-engine API
  version: 0.1.0
  description: API contract for qa-automation-engine.
servers:
  - url: http://localhost:8080
paths:
...
  /recurring-payments/v1/PN/recurring-payments:
    post:
      operationId: createPnRecurringPayment
      summary: Create PN recurring payment
      parameters:
        - $ref: '#/components/parameters/XFapiInteractionId'
        - $ref: '#/components/parameters/XJwsSignature'
        - $ref: '#/components/parameters/XIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentsRequest'
      responses:
        "201":
          description: Recurring payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentsResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /recurring-payments/v1/PJ/recurring-payments:
    post:
      operationId: createPjRecurringPayment
      summary: Create PJ recurring payment
      parameters:
        - $ref: '#/components/parameters/XFapiInteractionId'
        - $ref: '#/components/parameters/XJwsSignature'
        - $ref: '#/components/parameters/XIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentsRequest'
      responses:
        "201":
          description: Recurring payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentsResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /recurring-payments/v1/{participantType}/recurring-payments/{paymentId}:
    get:
      operationId: getRecurringPayment
      summary: Get recurring payment detail
      parameters:
        - $ref: '#/components/parameters/ParticipantType'
        - $ref: '#/components/parameters/PaymentId'
        - $ref: '#/components/parameters/XFapiInteractionId'
        - $ref: '#/components/parameters/XJwsSignatureOptional'
      responses:
        "200":
          description: Recurring payment detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDetailResponse'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

  # ── Variable Recurring Payments ───────────────────────────────────────────
...
components:
  parameters:
    XFapiInteractionId:
      name: x-fapi-interaction-id
      in: header
      required: true
      schema:
        type: string
    XJwsSignature:
      name: x-jws-signature
      in: header
      required: true
      schema:
        type: string
    XJwsSignatureOptional:
      name: x-jws-signature
      in: header
      required: false
      schema:
        type: string
    XIdempotencyKey:
      name: x-idempotency-key
      in: header
      required: true
      schema:
        type: string
    ParticipantType:
      name: participantType
      in: path
      required: true
      schema:
        type: string
        enum: [PN, PJ]
    PaymentId:
      name: paymentId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Payment not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Unprocessable entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    PaymentsRequest:
      type: object
      additionalProperties: true
    PaymentsResponse:
      type: object
      additionalProperties: true
    PaymentDetailResponse:
      type: object
      additionalProperties: true
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        traceId:
          type: string